feat: Add total in GET /api/v2/contests response
